+++ apps/montelo/rollup.config.cjs
@@ -1,8 +1,6 @@
 const resolve = require("@rollup/plugin-node-resolve").nodeResolve;
 const commonjs = require("@rollup/plugin-commonjs");
 const typescript = require("rollup-plugin-typescript2");
-const json = require("@rollup/plugin-json");
-const wasm = require("@rollup/plugin-wasm");
 
 /**
  * @type {import("rollup").RollupOptions[]}
@@ -14,7 +12,7 @@ module.exports = [
       dir: "dist/bundle.js",
       format: "commonjs",
     },
-    plugins: [typescript(), resolve(), commonjs(), json(), wasm()],
+    plugins: [typescript(), resolve(), commonjs()],
   },
   // Separate configuration for the CLI script
   {