+++ apps/montelo/src/Montelo.ts
@@ -3,7 +3,6 @@ import cuid from "cuid";
 
 import { MonteloClient } from "./MonteloClient";
 import { LogInput } from "./client";
-import { ExtendedMistral } from "./extended/ExtendedMistral";
 import { ExtendedOpenAI } from "./extended/ExtendedOpenAI";
 import { MonteloOptions, TraceParams } from "./types";
 
@@ -15,13 +14,11 @@ export class Montelo {
   private readonly constructorOptions: MonteloOptions | undefined;
   private readonly monteloClient: MonteloClient;
   public readonly openai: ExtendedOpenAI;
-  public readonly mistral: ExtendedMistral;
 
   constructor(options?: MonteloOptions) {
     this.constructorOptions = options;
     this.monteloClient = new MonteloClient(options?.montelo);
     this.openai = new ExtendedOpenAI(this.monteloClient, options?.openai);
-    this.mistral = new ExtendedMistral(this.monteloClient, options?.mistral);
   }
 
   public log(log: LogParams) {