+++ apps/montelo/src/extended/ExtendedOpenAI.ts
@@ -1,13 +1,17 @@
 import { LogSources } from "@montelo/db";
-import { TiktokenModel, encodingForModel } from "js-tiktoken";
 import OpenAI, { ClientOptions as OpenAIClientOptions } from "openai";
 import { APIPromise, RequestOptions } from "openai/core";
+import { Chat } from "openai/resources";
 import { ChatCompletionCreateParamsBase } from "openai/resources/chat/completions";
 import { Stream } from "openai/streaming";
 
+
+
 import { MonteloClient } from "../MonteloClient";
 import { LogInput } from "../client";
-import { MonteloLogExtend } from "./types";
+import { MonteloLogExtend, separateExtend } from "./types";
+
+import ChatCompletion = Chat.ChatCompletion;
 
 class ExtendedChatCompletions extends OpenAI.Chat.Completions {
   private monteloClient: MonteloClient;
@@ -37,14 +41,14 @@ class ExtendedChatCompletions extends OpenAI.Chat.Completions {
   ): APIPromise<OpenAI.ChatCompletion | Stream<OpenAI.ChatCompletionChunk>> {
     const startTime = new Date();
 
-    const { name, extra, ...bodyWithoutMonteloParams } = body;
-    const originalPromise = super.create(bodyWithoutMonteloParams, options);
+    const { base, extend } = separateExtend(body);
+    const originalPromise = super.create(base, options);
 
-    if ("stream" in body && body.stream) {
+    if ("stream" in base && base.stream) {
       // @ts-ignore
       return originalPromise.then((originalStream) => {
         // @ts-ignore
-        return this.augmentStream(originalStream, startTime, body);
+        return this.augmentStream(originalStream, startTime, base, extend);
       });
     } else {
       // @ts-ignore
@@ -52,22 +56,11 @@ class ExtendedChatCompletions extends OpenAI.Chat.Completions {
         const endTime = new Date();
         const duration = ((endTime.getTime() - startTime.getTime()) / 1000).toFixed(2);
 
-        // @ts-ignore
-        void this.logToDatabase(
-          body,
-          {
-            model: body.model,
-            // @ts-ignore
-            content: data.choices[0],
-            // @ts-ignore
-            usage: data.usage,
-          },
-          {
-            startTime: startTime.toISOString(),
-            endTime: endTime.toISOString(),
-            duration: parseFloat(duration),
-          },
-        );
+        void this.logToDatabase(base, extend, data as OpenAI.ChatCompletion, {
+          startTime: startTime.toISOString(),
+          endTime: endTime.toISOString(),
+          duration: parseFloat(duration),
+        });
         return data;
       });
     }
@@ -76,7 +69,8 @@ class ExtendedChatCompletions extends OpenAI.Chat.Completions {
   private augmentStream(
     originalStream: Stream<OpenAI.ChatCompletionChunk>,
     startTime: Date,
-    body: ChatCompletionCreateParamsBase & MonteloLogExtend,
+    base: ChatCompletionCreateParamsBase,
+    extend: MonteloLogExtend,
   ): Stream<OpenAI.ChatCompletionChunk> {
     const self = this;
 
@@ -99,44 +93,31 @@ class ExtendedChatCompletions extends OpenAI.Chat.Completions {
         return acc;
       }, "");
 
-      function getNumberOfTokens(text: string, modelIdentifier: TiktokenModel): number {
-        // Get the encoding for the specified model
-        const enc = encodingForModel(modelIdentifier);
-
-        // Encode the text
-        const encodedText = enc.encode(text);
-
-        // Return the number of tokens
-        return encodedText.length;
-      }
-
-      const concattedInput = body.messages.reduce((accum, curr) => {
-        return accum + curr.content;
-      }, "");
-      const inputTokens = getNumberOfTokens(concattedInput, body.model as TiktokenModel);
-      const outputTokens = getNumberOfTokens(finalContent, body.model as TiktokenModel);
-      const totalTokens = inputTokens + outputTokens;
-
-      // token counts
-      const usage = {
-        prompt_tokens: inputTokens,
-        completion_tokens: outputTokens,
-        total_tokens: totalTokens,
+      const data: ChatCompletion = {
+        id: chunks[0].id,
+        system_fingerprint: chunks[0].system_fingerprint,
+        created: chunks[0].created,
+        model: chunks[0].model,
+        usage: undefined,
+        object: "chat.completion",
+        choices: [
+          {
+            index: chunks[0].choices[0].index,
+            logprobs: chunks[0].choices[0].logprobs || null,
+            finish_reason: chunks[0].choices[0].finish_reason || "stop",
+            message: {
+              content: finalContent,
+              role: "assistant",
+            },
+          },
+        ],
       };
 
-      await self.logToDatabase(
-        body,
-        {
-          model: body.model,
-          content: { output: finalContent },
-          usage,
-        },
-        {
-          startTime: startTime.toISOString(),
-          endTime: endTime.toISOString(),
-          duration: parseFloat(duration),
-        },
-      );
+      await self.logToDatabase(base, extend, data, {
+        startTime: startTime.toISOString(),
+        endTime: endTime.toISOString(),
+        duration: parseFloat(duration),
+      });
     }
 
     // Return a new Stream instance wrapping the augmented iterator
@@ -145,16 +126,9 @@ class ExtendedChatCompletions extends OpenAI.Chat.Completions {
   }
 
   private async logToDatabase(
-    input: ChatCompletionCreateParamsBase & MonteloLogExtend,
-    output: {
-      model: string;
-      content: Record<string, any>;
-      usage: {
-        prompt_tokens: number;
-        completion_tokens: number;
-        total_tokens: number;
-      };
-    },
+    base: ChatCompletionCreateParamsBase,
+    extend: MonteloLogExtend,
+    output: OpenAI.ChatCompletion,
     time: {
       startTime: string;
       endTime: string;
@@ -163,15 +137,18 @@ class ExtendedChatCompletions extends OpenAI.Chat.Completions {
   ): Promise<void> {
     const log: LogInput = {
       ...time,
+      ...extend,
+      ...(output.usage && {
+        tokens: {
+          inputTokens: output.usage.prompt_tokens,
+          outputTokens: output.usage.completion_tokens,
+          totalTokens: output.usage.total_tokens,
+        },
+      }),
       source: LogSources.OPENAI,
-      name: input.name,
       model: output.model,
-      inputTokens: output.usage?.prompt_tokens,
-      outputTokens: output.usage?.completion_tokens,
-      totalTokens: output.usage?.total_tokens,
-      input: input.messages,
-      output: output.content,
-      extra: input.extra,
+      input: base,
+      output,
     };
     await this.monteloClient.createLog(log);
   }