+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @montelo/config
 
+## 1.1.49
+
+### Patch Changes
+
+- y
+
 ## 1.1.48
 
 ### Patch Changes