+++ packages/promptfile/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @montelo/promptfile
 
+## 1.1.49
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/common@1.1.49
+
 ## 1.1.48
 
 ### Patch Changes